refactor: 이벤트 채널 업데이트
