+++ js/actions/index.js
@@ -1,0 +1,5 @@
+export const ACTION = {
+  INIT: 'init',
+  VIEW_INIT: 'viewInit',
+  ADD_USER: 'addUser',
+};
